counting from 100 to 1 through while loop.c
In this program the user is asked to print the counting backwards from 100 to 1.
here, the initial value of the variable i is taken as 100, printed and decreased with each iteration the loop takes place. 
the loop goes by:
100
99
98
97
.
.
.
.
1
we have used the unary decrement operator (--) here to decrease the value of the variable i with each iteration.